+++ src/matrix/update/traits.rs
@@ -45,7 +45,7 @@ pub trait InPlaceUpdates<T:Number> : Shape<T>+MatrixBuffer<T> {
 
    /// Sets all the entries on a row to a particular value
     fn set_row(&mut self, 
-        r :  uint, 
+        r :  usize, 
         value : T
         )-> &mut Self {
         debug_assert! (r  < self.num_rows());
@@ -61,7 +61,7 @@ pub trait InPlaceUpdates<T:Number> : Shape<T>+MatrixBuffer<T> {
 
    /// Sets all the entries on a column to a particular value
     fn set_col(&mut self, 
-        c :  uint, 
+        c :  usize, 
         value : T
         )-> &mut Self {
         debug_assert! (c  < self.num_cols());
@@ -76,14 +76,14 @@ pub trait InPlaceUpdates<T:Number> : Shape<T>+MatrixBuffer<T> {
 
    /// Sets all the entries on a block to a particular value
     fn set_block(&mut self, 
-        start_row : int, 
-        start_col : int , 
-        num_rows: uint, 
-        num_cols : uint,
+        start_row : isize, 
+        start_col : isize , 
+        num_rows: usize, 
+        num_cols : usize,
         value : T
         )-> &mut Self {
-        let r = mod_n(start_row, self.num_rows() as int);        
-        let c = mod_n(start_col, self.num_cols() as int);
+        let r = mod_n(start_row, self.num_rows() as isize);        
+        let c = mod_n(start_col, self.num_cols() as isize);
         assert!(num_rows < self.num_rows());
         assert!(num_cols < self.num_cols());
         let p = self.as_mut_ptr();
@@ -104,13 +104,13 @@ pub trait InPlaceUpdates<T:Number> : Shape<T>+MatrixBuffer<T> {
      /// Divide the matrix by a scalar
     fn div_scalar(&mut self, rhs: T) -> SRResult<&mut Self>;
     /// Scales a specific row in lower triangular part 
-    fn scale_row_lt(&mut self, r :  uint, scale_factor : T)-> &mut Self;
+    fn scale_row_lt(&mut self, r :  usize, scale_factor : T)-> &mut Self;
     /// Scales a specific column in lower triangular part
-    fn scale_col_lt(&mut self, c :  uint, scale_factor : T)-> &mut Self;
+    fn scale_col_lt(&mut self, c :  usize, scale_factor : T)-> &mut Self;
     /// Scales a specific row in upper triangular part
-    fn scale_row_ut(&mut self, r :  uint, scale_factor : T)-> &mut Self;
+    fn scale_row_ut(&mut self, r :  usize, scale_factor : T)-> &mut Self;
     /// Scales a specific column in upper triangular part
-    fn scale_col_ut(&mut self, c :  uint, scale_factor : T)-> &mut Self;
+    fn scale_col_ut(&mut self, c :  usize, scale_factor : T)-> &mut Self;
     /// Scale all rows as per the scale factors
     fn scale_rows(&mut self, scale_factors : &Matrix<T>)-> &mut Self;
     /// Scale all columns as per the scale factors
