+++ src/util/memory.rs
@@ -11,7 +11,7 @@ Private helper functions follow
 
 /// Deallocates a buffer of memory
 #[inline]
-pub unsafe fn dealloc<T>(ptr: *mut T, len: uint) {
+pub unsafe fn dealloc<T>(ptr: *mut T, len: usize) {
     if mem::size_of::<T>() != 0 {
         deallocate(ptr as *mut u8,
                    len * mem::size_of::<T>(),
